# sync: reduce contention in `Notify`