--- a/benches/Cargo.toml
+++ b/benches/Cargo.toml
@@ -42,6 +42,10 @@ name = "rt_multi_threaded"
 path = "rt_multi_threaded.rs"
 harness = false
 
+[[bench]]
+name = "sync_notify"
+path = "sync_notify.rs"
+harness = false
 
 [[bench]]
 name = "sync_rwlock"
@@ -67,3 +71,4 @@ harness = false
 name = "copy"
 path = "copy.rs"
 harness = false
+