@Darksonn 2023-02-26
This PR is rather large. Will take me a bit until I can review it.