@satakuma 2023-02-26
Sure, no problem. I've pushed an update which slightly reduces the size of this change. I can write down some key points to make the review easier.

Adding the atomic variable is quite straightforward, it boils down to adding atomic stores to `notify_{one,waiters}` functions and an atomic load to `poll_notified`. The reason why the diff ended up being so large is that I had to change the existing code to be correct aliasing-wise regarding the atomic field.

Previously, a `Waiter` lived inside an `UnsafeCell` and could be mutably borrowed after acquiring the lock which guards the waitlist. This PR avoids mutable borrows of `Waiter` and instead wraps its only mutable field in an `UnsafeCell` to allow mutation. *Technically*, there is one, indirect mutable borrow of `Waiter` coming from poll's `Pin<&mut Notified>`, but that's a known issue of self-referential structs and this is safe because `Waiter` is `!Unpin`.

Waiters in `BatchSemaphore` also use an atomic field for tracking the number of permits, and the code there is organized similarly to this PR. If this PR gets merged, I would like to also investigate if similar optimization can be applied in the IO driver.