@satakuma 2023-03-17
I've added some comments to better document the safety invariants. I've also moved the atomic variable into a wrapper type to use an enum instead of constants.